clean sidemenu
