+++ packages/core/src/extensions/SideMenu/SideMenuPlugin.ts
@@ -1,11 +1,7 @@
 import { PluginView } from "@tiptap/pm/state";
-import { Node } from "prosemirror-model";
-import { NodeSelection, Plugin, PluginKey, Selection } from "prosemirror-state";
+import { Plugin, PluginKey } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
-import * as pmView from "prosemirror-view";
 
-import { createExternalHTMLExporter } from "../../api/exporters/html/externalHTMLExporter";
-import { cleanHTMLToMarkdown } from "../../api/exporters/markdown/markdownExporter";
 import { getBlockInfoFromPos } from "../../api/getBlockInfoFromPos";
 import { Block } from "../../blocks/defaultBlocks";
 import type { BlockNoteEditor } from "../../editor/BlockNoteEditor";
@@ -13,9 +9,7 @@ import { UiElementPosition } from "../../extensions-shared/UiElementPosition";
 import { BlockSchema, InlineContentSchema, StyleSchema } from "../../schema";
 import { EventEmitter } from "../../util/EventEmitter";
 import { initializeESMDependencies } from "../../util/esmDependencies";
-import { MultipleNodeSelection } from "./MultipleNodeSelection";
-
-let dragImageElement: Element | undefined;
+import { dragStart, unsetDragImage } from "./dragging";
 
 export type SideMenuState<
   BSchema extends BlockSchema,
@@ -26,230 +20,9 @@ export type SideMenuState<
   block: Block<BSchema, I, S>;
 };
 
-export function getDraggableBlockFromElement(
-  element: Element,
-  view: EditorView
-) {
-  while (
-    element &&
-    element.parentElement &&
-    element.parentElement !== view.dom &&
-    element.getAttribute?.("data-node-type") !== "blockContainer"
-  ) {
-    element = element.parentElement;
-  }
-  if (element.getAttribute?.("data-node-type") !== "blockContainer") {
-    return undefined;
-  }
-  return { node: element as HTMLElement, id: element.getAttribute("data-id")! };
-}
-
-function blockPositionFromElement(element: Element, view: EditorView) {
-  const block = getDraggableBlockFromElement(element, view);
-
-  if (block && block.node.nodeType === 1) {
-    // TODO: this uses undocumented PM APIs? do we need this / let's add docs?
-    const docView = (view as any).docView;
-    const desc = docView.nearestDesc(block.node, true);
-    if (!desc || desc === docView) {
-      return null;
-    }
-    return desc.posBefore;
-  }
-  return null;
-}
-
-function blockPositionsFromSelection(selection: Selection, doc: Node) {
-  // Absolute positions just before the first block spanned by the selection, and just after the last block. Having the
-  // selection start and end just before and just after the target blocks ensures no whitespace/line breaks are left
-  // behind after dragging & dropping them.
-  let beforeFirstBlockPos: number;
-  let afterLastBlockPos: number;
-
-  // Even the user starts dragging blocks but drops them in the same place, the selection will still be moved just
-  // before & just after the blocks spanned by the selection, and therefore doesn't need to change if they try to drag
-  // the same blocks again. If this happens, the anchor & head move out of the block content node they were originally
-  // in. If the anchor should update but the head shouldn't and vice versa, it means the user selection is outside a
-  // block content node, which should never happen.
-  const selectionStartInBlockContent =
-    doc.resolve(selection.from).node().type.spec.group === "blockContent";
-  const selectionEndInBlockContent =
-    doc.resolve(selection.to).node().type.spec.group === "blockContent";
-
-  // Ensures that entire outermost nodes are selected if the selection spans multiple nesting levels.
-  const minDepth = Math.min(selection.$anchor.depth, selection.$head.depth);
-
-  if (selectionStartInBlockContent && selectionEndInBlockContent) {
-    // Absolute positions at the start of the first block in the selection and at the end of the last block. User
-    // selections will always start and end in block content nodes, but we want the start and end positions of their
-    // parent block nodes, which is why minDepth - 1 is used.
-    const startFirstBlockPos = selection.$from.start(minDepth - 1);
-    const endLastBlockPos = selection.$to.end(minDepth - 1);
-
-    // Shifting start and end positions by one moves them just outside the first and last selected blocks.
-    beforeFirstBlockPos = doc.resolve(startFirstBlockPos - 1).pos;
-    afterLastBlockPos = doc.resolve(endLastBlockPos + 1).pos;
-  } else {
-    beforeFirstBlockPos = selection.from;
-    afterLastBlockPos = selection.to;
-  }
-
-  return { from: beforeFirstBlockPos, to: afterLastBlockPos };
-}
-
-function setDragImage(view: EditorView, from: number, to = from) {
-  if (from === to) {
-    // Moves to position to be just after the first (and only) selected block.
-    to += view.state.doc.resolve(from + 1).node().nodeSize;
-  }
-
-  // Parent element is cloned to remove all unselected children without affecting the editor content.
-  const parentClone = view.domAtPos(from).node.cloneNode(true) as Element;
-  const parent = view.domAtPos(from).node as Element;
-
-  const getElementIndex = (parentElement: Element, targetElement: Element) =>
-    Array.prototype.indexOf.call(parentElement.children, targetElement);
-
-  const firstSelectedBlockIndex = getElementIndex(
-    parent,
-    // Expects from position to be just before the first selected block.
-    view.domAtPos(from + 1).node.parentElement!
-  );
-  const lastSelectedBlockIndex = getElementIndex(
-    parent,
-    // Expects to position to be just after the last selected block.
-    view.domAtPos(to - 1).node.parentElement!
-  );
-
-  for (let i = parent.childElementCount - 1; i >= 0; i--) {
-    if (i > lastSelectedBlockIndex || i < firstSelectedBlockIndex) {
-      parentClone.removeChild(parentClone.children[i]);
-    }
-  }
-
-  // dataTransfer.setDragImage(element) only works if element is attached to the DOM.
-  unsetDragImage(view.root);
-  dragImageElement = parentClone;
-
-  // TODO: This is hacky, need a better way of assigning classes to the editor so that they can also be applied to the
-  //  drag preview.
-  const classes = view.dom.className.split(" ");
-  const inheritedClasses = classes
-    .filter(
-      (className) =>
-        className !== "ProseMirror" &&
-        className !== "bn-root" &&
-        className !== "bn-editor"
-    )
-    .join(" ");
-
-  dragImageElement.className =
-    dragImageElement.className + " bn-drag-preview " + inheritedClasses;
-
-  if (view.root instanceof ShadowRoot) {
-    view.root.appendChild(dragImageElement);
-  } else {
-    view.root.body.appendChild(dragImageElement);
-  }
-}
-
-function unsetDragImage(rootEl: Document | ShadowRoot) {
-  if (dragImageElement !== undefined) {
-    if (rootEl instanceof ShadowRoot) {
-      rootEl.removeChild(dragImageElement);
-    } else {
-      rootEl.body.removeChild(dragImageElement);
-    }
-
-    dragImageElement = undefined;
-  }
-}
-
-function dragStart<
-  BSchema extends BlockSchema,
-  I extends InlineContentSchema,
-  S extends StyleSchema
->(
-  e: { dataTransfer: DataTransfer | null; clientY: number },
-  editor: BlockNoteEditor<BSchema, I, S>
-) {
-  if (!e.dataTransfer) {
-    return;
-  }
-
-  const view = editor.prosemirrorView;
-
-  const editorBoundingBox = view.dom.getBoundingClientRect();
-
-  const coords = {
-    left: editorBoundingBox.left + editorBoundingBox.width / 2, // take middle of editor
-    top: e.clientY,
-  };
-
-  const elements = view.root.elementsFromPoint(coords.left, coords.top);
-  let blockEl = undefined;
-
-  for (const element of elements) {
-    if (view.dom.contains(element)) {
-      blockEl = getDraggableBlockFromElement(element, view);
-      break;
-    }
-  }
-
-  if (!blockEl) {
-    return;
-  }
-
-  const pos = blockPositionFromElement(blockEl.node, view);
-  if (pos != null) {
-    const selection = view.state.selection;
-    const doc = view.state.doc;
-
-    const { from, to } = blockPositionsFromSelection(selection, doc);
-
-    const draggedBlockInSelection = from <= pos && pos < to;
-    const multipleBlocksSelected =
-      selection.$anchor.node() !== selection.$head.node() ||
-      selection instanceof MultipleNodeSelection;
-
-    if (draggedBlockInSelection && multipleBlocksSelected) {
-      view.dispatch(
-        view.state.tr.setSelection(MultipleNodeSelection.create(doc, from, to))
-      );
-      setDragImage(view, from, to);
-    } else {
-      view.dispatch(
-        view.state.tr.setSelection(NodeSelection.create(view.state.doc, pos))
-      );
-      setDragImage(view, pos);
-    }
-
-    const selectedSlice = view.state.selection.content();
-    const schema = editor.pmSchema;
-
-    const clipboardHML = (pmView as any).__serializeForClipboard(
-      view,
-      selectedSlice
-    ).dom.innerHTML;
-
-    const externalHTMLExporter = createExternalHTMLExporter(schema, editor);
-    const externalHTML = externalHTMLExporter.exportProseMirrorFragment(
-      selectedSlice.content,
-      {}
-    );
-
-    const plainText = cleanHTMLToMarkdown(externalHTML);
-
-    e.dataTransfer.clearData();
-    e.dataTransfer.setData("blocknote/html", clipboardHML);
-    e.dataTransfer.setData("text/html", externalHTML);
-    e.dataTransfer.setData("text/plain", plainText);
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setDragImage(dragImageElement!, 0, 0);
-    view.dragging = { slice: selectedSlice, move: true };
-  }
-}
-
+/**
+ * With the sidemenu plugin we can position a menu next to a hovered block.
+ */
 export class SideMenuView<
   BSchema extends BlockSchema,
   I extends InlineContentSchema,
@@ -259,7 +32,6 @@ export class SideMenuView<
   public state?: SideMenuState<BSchema, I, S>;
   public readonly emitUpdate: (state: SideMenuState<BSchema, I, S>) => void;
 
-  private needUpdate = false;
   private mousePos: { x: number; y: number } | undefined;
 
   // When true, the drag handle with be anchored at the same level as root elements
@@ -271,7 +43,7 @@ export class SideMenuView<
   private hoveredBlock: HTMLElement | undefined;
 
   // Used to check if currently dragged content comes from this editor instance.
-  public isDragging = false;
+  public isDragging = false; // TODO
 
   public menuFrozen = false;
 
@@ -330,7 +102,12 @@ export class SideMenuView<
     this.pmView.root.addEventListener("scroll", this.onScroll, true);
   }
 
-  updateState = () => {
+  updateState = (state: SideMenuState<BSchema, I, S>) => {
+    this.state = state;
+    this.emitUpdate(this.state);
+  };
+
+  updateStateFromMousePos = () => {
     if (this.menuFrozen || !this.mousePos) {
       return;
     }
@@ -372,7 +149,7 @@ export class SideMenuView<
     if (!block || !this.editor.isEditable) {
       if (this.state?.show) {
         this.state.show = false;
-        this.needUpdate = true;
+        this.updateState(this.state);
       }
 
       return;
@@ -396,11 +173,13 @@ export class SideMenuView<
       return;
     }
 
+    // TODO: needed?
+
     // Shows or updates elements.
     if (this.editor.isEditable) {
       const blockContentBoundingBox = blockContent.getBoundingClientRect();
 
-      this.state = {
+      this.updateState({
         show: true,
         referencePos: new DOMRect(
           this.horizontalPosAnchoredAtRoot
@@ -413,13 +192,14 @@ export class SideMenuView<
         block: this.editor.getBlock(
           this.hoveredBlock!.getAttribute("data-id")!
         )!,
-      };
-      this.needUpdate = true;
+      });
     }
   };
 
   /**
    * Sets isDragging when dragging text.
+   *
+   * TODO: clean up naming confusing
    */
   onDragStart = () => {
     this.isDragging = true;
@@ -504,6 +284,7 @@ export class SideMenuView<
     }
   };
 
+  // TODO
   onMouseDown = () => {
     if (this.state && this.state.show && this.menuFrozen) {
       this.menuFrozen = false;
@@ -528,7 +309,8 @@ export class SideMenuView<
       this.mousePos.y > editorOuterBoundingBox.top &&
       this.mousePos.y < editorOuterBoundingBox.bottom;
 
-    const editorWrapper = this.pmView.dom.parentElement!;
+    // TODO: remove parentElement, but then we need to remove padding from boundingbox or find a different solution
+    const editorWrapper = this.pmView.dom!.parentElement!;
 
     // Doesn't update if the mouse hovers an element that's over the editor but
     // isn't a part of it or the side menu.
@@ -552,14 +334,10 @@ export class SideMenuView<
       return;
     }
 
-    this.updateState();
-
-    if (this.needUpdate) {
-      this.emitUpdate(this.state!);
-      this.needUpdate = false;
-    }
+    this.updateStateFromMousePos();
   };
 
+  // TODO: needed?
   onScroll = () => {
     if (this.state?.show && this.hoveredBlock?.firstChild) {
       const blockContent = this.hoveredBlock.firstChild as HTMLElement;
@@ -585,14 +363,7 @@ export class SideMenuView<
   // would otherwise not update the side menu, and so clicking the button again
   // would attempt to remove the same block again, causing an error.
   update() {
-    const prevBlockId = this.state?.block.id;
-
-    this.updateState();
-
-    if (this.needUpdate && this.state && prevBlockId !== this.state.block.id) {
-      this.emitUpdate(this.state);
-      this.needUpdate = false;
-    }
+    this.updateStateFromMousePos();
   }
 
   destroy() {
@@ -742,4 +513,22 @@ export class SideMenuProsemirrorPlugin<
     this.view!.state!.show = false;
     this.view!.emitUpdate(this.view!.state!);
   };
+}
+
+export function getDraggableBlockFromElement(
+  element: Element,
+  view: EditorView
+) {
+  while (
+    element &&
+    element.parentElement &&
+    element.parentElement !== view.dom &&
+    element.getAttribute?.("data-node-type") !== "blockContainer"
+  ) {
+    element = element.parentElement;
+  }
+  if (element.getAttribute?.("data-node-type") !== "blockContainer") {
+    return undefined;
+  }
+  return { node: element as HTMLElement, id: element.getAttribute("data-id")! };
 }
